Don't delete S3 bucket files if the build faulted so we can debug (#120)
Don't cleanup files if job faulted so it can be recreate build

+++ src/SIL.Machine.AspNetCore/Services/NmtPostprocessBuildJob.cs
@@ -58,7 +58,8 @@ public class NmtPostprocessBuildJob : HangfireBuildJob<(int, double)>
 
         try
         {
-            await _sharedFileService.DeleteAsync($"builds/{buildId}/");
+            if (completionStatus is not JobCompletionStatus.Faulted)
+                await _sharedFileService.DeleteAsync($"builds/{buildId}/");
         }
         catch (Exception e)
         {
